docs(storybook): apply the deletion rule to the retained motion and state stories
The earlier cut used the wrong test. It asked "is this Maka's own?" and kept
whatever answered yes. The question that matters is whether Astryx already owns
the domain — if it does, a Maka catalogue of the same thing is a parallel
vocabulary being documented as the contract, which is exactly what the six
deleted token stories were.

Animation Catalog failed that test. Its DurationScale and EasingScale presented
--duration-quick/base/emphasized/large and the four --ease-* tokens as the motion
scale, while the Astryx theme already ships --duration-fast/medium/slow. Both are
gone. What remains is the spinner and the streaming shimmer — load-bearing, not
decorative, and they outlive the token rename — so the file is now
Design System/Functional Motion, which is what it actually shows.

Interaction States failed it too, in the part nobody flagged: ButtonStates,
NeutralButtonStates and SolidButtonStates are state matrices for Astryx's Button,
the same component whose Primitives/Button story this branch already deleted.

ListRowStates stays, and it is the reason the file stays: it compares hover and
focus across two DIFFERENT row components — Astryx's side-nav item and its list
item — as the sidebar composes them. session-list-panel.stories.tsx owns
SessionListPanel's own states; nothing else covers the seam between the two. The
file is now Design System/Composite Row States.

Design System/* is left holding Icons and Functional Motion — the Maka icon set,
the brand mark, and two functional animations — plus that one cross-component
seam. Nothing in it catalogues a token scale any more.
